Move mutable matrix files into basic-mutable-matrices and mutable-matrices
