+++ M2/Macaulay2/e/basic-mutable-matrices/mat-util.hpp
