+++ M2/Macaulay2/e/basic-mutable-matrices/smat.hpp
