+++ M2/Macaulay2/e/basic-rings/aring-glue.hpp
@@ -7,7 +7,7 @@
 #include "basic-rings/aring-translate.hpp"
 #include "ring.hpp"
 
-#include "mutablemat.hpp"
+#include "mutable-matrices/mutablemat.hpp"
 
 static const bool displayArithmeticCalls = false;
