+++ M2/Macaulay2/e/debug.cpp
@@ -6,7 +6,7 @@
 #include "res-a1-poly.hpp"
 #include "res-a0-poly.hpp"
 #include "hermite.hpp"
-#include "mat.hpp"
+#include "basic-mutable-matrices/mat.hpp"
 #include "monideal.hpp"
 
 void showint(mpz_srcptr a)
