+++ M2/Macaulay2/e/f4/f4.cpp
@@ -2,7 +2,7 @@
 
 #include "f4/f4.hpp"
 #include "freemod.hpp"                 // for FreeModule
-#include "mat.hpp"                     // for MutableMatrix
+#include "basic-mutable-matrices/mat.hpp"                     // for MutableMatrix
 #include "text-io.hpp"                 // for emit
 #include "buffer.hpp"                  // for buffer
 #include "error.h"                     // for error
