+++ M2/Macaulay2/e/interface/matrix.cpp
@@ -7,17 +7,17 @@
 #include "NAG.hpp" // TODO: can this be removed?
 #include "SLP-defs.hpp"
 #include "buffer.hpp"
-#include "dmat.hpp"
+#include "basic-mutable-matrices/dmat.hpp"
 #include "error.h"
 #include "exceptions.hpp"
 #include "freemod.hpp"
 #include "interface/NAG.h"
 #include "interface/gmp-util.h"
 #include "interface/monoid.h"
-#include "mat.hpp"
+#include "basic-mutable-matrices/mat.hpp"
 #include "matrix-con.hpp"
 #include "matrix.hpp"
-#include "mutablemat-defs.hpp"
+#include "mutable-matrices/mutablemat-defs.hpp"
 #include "relem.hpp"
 #include "ring.hpp"
 #include "ringelem.hpp"
