+++ M2/Macaulay2/e/interface/mutable-matrix.cpp
@@ -15,7 +15,7 @@
 #include "fractionfreeLU.hpp"
 #include "interface/gmp-util.h"
 #include "interface/random.h"
-#include "mat.hpp"
+#include "basic-mutable-matrices/mat.hpp"
 #include "matrix.hpp"
 #include "relem.hpp"
 #include "ring.hpp"
