+++ M2/Macaulay2/e/interface/ringmap.cpp
@@ -4,7 +4,7 @@
 #include "error.h"
 #include "exceptions.hpp"
 #include "freemod.hpp"
-#include "mat.hpp"
+#include "basic-mutable-matrices/mat.hpp"
 #include "matrix.hpp"
 #include "ringmap.hpp"
