+++ M2/Macaulay2/e/ntl-interface.cpp
@@ -1,7 +1,7 @@
 // Copyright 2005, Michael Stillman
 
 #include "ntl-interface.hpp"
-#include "mat.hpp"
+#include "basic-mutable-matrices/mat.hpp"
 
 NTL::mat_ZZ *mutableMatrix_to_NTL_mat_ZZ(const MutableMatrix *M)
 {
